Fix tests for event picker table
