+++ tnoodle-ui/src/test/mock/util.test.mock.ts
@@ -1,0 +1,6 @@
+export const axiosResponse = {
+    status: 200,
+    statusText: "OK",
+    config: {},
+    headers: {},
+};
